requirements.txt
astroid==1.3.6
Django==1.8
django-autofixture==0.10.0
django-braces==1.8.0
logilab-common==0.63.2
psycopg2==2.6
pylint==1.4.3
selenium==2.45.0
six==1.9.0
